backend/requirements.txt
fastapi[standard]
uvicorn[standard]
requests
beautifulsoup4
lxml
boto3
python-dotenv
serpapi
google-search-results
sqlalchemy
psycopg2-binary